requirements.txt
# TODO:

flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
smtplib-ssl
PyGithub>=1.59.0